Avoid going back and forth between signed and unsigned in the same expression with some casts and an unsigned constant.  This quiets GCC warnings, especially -Wconversion, IIRC.

+++ src/H5time_private.h
@@ -103,7 +103,7 @@
 			(vsp)->tv_nsec += 1000000000L;			\
 		}							\
 	} while (/* CONSTCOND */ 0)
-#define timespec2ns(x) (((uint64_t)(x)->tv_sec) * 1000000000L + (x)->tv_nsec)
+#define timespec2ns(x) (((uint64_t)(x)->tv_sec) * 1000000000UL + (uint64_t)(x)->tv_nsec)
 #endif
 
 #endif /* _H5time_private_H */
